Mocks update
Moved all the mocks to separate files. All the mocks will be stored in
 `common/mocks/*`
