+++ applications/ExecutionManager/test/application_handler_test.cpp
@@ -1,5 +1,5 @@
 #include "application_handler.hpp"
-#include <mocks.hpp>
+#include <mocks/os_interface_mock.hpp>
 
 #include <memory>
