+++ common/include/mocks/execution_manager_client_mock.hpp
@@ -1,0 +1,17 @@
+#ifndef EXECUTIONMANAGERCLIENTMOCK
+#define EXECUTIONMANAGERCLIENTMOCK
+
+#include "i_execution_manager_client.hpp"
+
+#include "gmock/gmock.h"
+
+using StateError = ::MachineStateManagement::StateError;
+
+class ExecutionManagerClientMock :
+  public ExecutionManagerClient::IExecutionManagerClient
+{
+public:
+  MOCK_METHOD(void, confirm, (StateError status));
+};
+
+#endif // EXECUTIONMANAGERCLIENTMOCK
