+++ common/include/mocks/i_state_mock.hpp
@@ -1,0 +1,16 @@
+#ifndef ISTATEMOCK
+#define ISTATEMOCK
+
+#include "i_state.hpp"
+
+#include "gmock/gmock.h"
+
+class IStateMock : public api::IState
+{
+public:
+    IStateMock() = default;
+    MOCK_METHOD(void, enter, ());
+    MOCK_METHOD(void, leave, (), (const));
+};
+
+#endif
